完成C端再来一单功能

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -64,5 +64,5 @@ public interface OrderMapper {
      * @return
      */
     @Select("select * from orders where id=#{id}")
-    Orders selectAllById(Integer id);
+    Orders selectAllById(Long id);
 }
